Add initial models, migrations, and templates
This commit adds the initial setup of models, migrations, and templates for the core and user modules. It also includes a service for calculating results from an Excel file. The models created cover sports, sports schools, average values standards and weighting factors. Migrations correspond to the created models. A template for the admin page was added as well.