Update model primary key and Django version in migrations
The primary key for models in Django migrations was changed from BigAutoField to UUIDField. In addition, the version of Django used to generate migrations was updated from 4.2.4 to 5.0. Some models have also been amended to inherit from the BaseModel class. Lastly, the import style used in the task.py file has been simplified for cleaner code.