Integrate 'drf_api_logger' into sport_connect_api
Added 'drf_api_logger' to installed apps and middleware in the settings.py file for sport_connect_api and added corresponding modules to the requirements.txt. This enhances the logging capabilities of the API, providing detailed logs for API requests and responses. The new settings will enable database storage for these logs.